KDKPAS-3101 Temporarily disable venv tests from pipeline
In this same ticket we made a patch to the warcio RPM that we've
packaged ourselves, but we haven't forked the warcio source code to our
own gitlab since we don't really want to upkeep it. So there isn't a
python package version of this patch that we could use here. Due to this
the venv tests in this repo fail at the moment. We decided to just
disable the venv tests at the moment until we can get these changes to
warcio upstream.